+++ packages/client/base/src/services/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./APIErrorService";
+export * from "./BaseCrossmintService";