+++ packages/client/base/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @crossmint/client-sdk-base
 
+## 1.2.3
+
+### Patch Changes
+
+-   Updated dependencies [8f56ddf]
+    -   @crossmint/common-sdk-base@0.1.2
+
 ## 1.2.2
 
 ### Patch Changes